requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
DateTime==4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-crispy-forms==1.14.0
django-import-export==2.8.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
MarkupPy==1.14
middleware==1.2.3
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
